jog #3 - recurring tasks
- changed add_task.php to return an object rather than array
- properly named variables
- used template literal string

Regarding your comment about the checkbox variable, I'm having trouble calling it in a different function from where it was initialized (the error says that the checkbox variable was not defined). I tried setting it as a global variable on the first line of code, outside of any function, but when I do that it seems to return null when the variable is used in any of my functions.

I just initialized the variable in every function that I used it in. Is there a way to fix the errors I mentioned above?